core(api): working on o_server_messages_migration integration
To satisfy the need of API sorted and limited number of messages
fetches in an API call I've created a new table within scylla, called
o_server_messages_migration with clustering key on datetime. This
allows for fetching sorted messages based on time... Moreover
it also allows for users to fetch only a number of messages, and
to offset the selection.