core(ODIVE-1): this fixes ODIVE-1 vulnerability impacting user tokens
ODIVE-1 (Online Division Internal Vurnerabilities and Exposures No.1)
is a vulnerability affecting user tokens, where a DB Admin would
have automatic access to any OD account by using the un-encrypted
tokens in the DB. This commit ensures that all tokens are properly
encrypted before being stored in the DB.

In addition to this we now have a different table for storing
tokens, a username-token-datetime database, this is so users
can have multiple devices active on a single account. This was
an issue that arised in the testing of the last od-client update.
It's now fixed.