core(spell_caster): auth routine implementation for od_spell_caster
This will be needed for the od-spell-caster implementation.
The current idea for auth-ing users is that we store
a pair of secrets (key & spell) in scylla, the spell-caster
gets access to both of them and it sends the user the key
on login. The user then uses the key (together with their
token) to login into the API and retrieve the spell.
Finally the user sends the spell to the spell-caster and
if it matches what he has, then the user gets logged in.
Here is a simple ilustration:

user ---> spell-caster ----> API
API ----(k,s)---> spell-caster ----(k)---> user
user ---(k)--->API----(s)---->user
user ---(s)----> spell-caster