core(api): added moka::future::Cache for token caching
Caching is kinda cool. You can hold key-value pairs (in this
case username-token pairs) in memory so that you can check
a token faster then by calling the Database each time.

This commit adds such caching. From now on, whenever the
API checks for a token, it first checks the in memory
moka cache instance (way faster then checking Scylla),
and if it doesn't find anything it checks Scylla (as
it did before this change was introduced).

I've set the maximum entry count to one thousand, this
can be debated: whether one thousand is too much for
one API instance.