core: [ODIVE-5] The API should use stronger password hashing
> [!WARNING]
> DO NOT MERGE UNLESS ENOUGH MEASURES HAVE BEEN TAKEN TO MIGRATE
> SCYLLA IN PREPARATION OF THIS CHANGE.

Currently we use SHA-512 for password hashing which is a practice
that's... rather seldomly put in practice by bigger projects. For good
reason. SHA is fast... too fast... to the point that it allows for
timely brute forcing of passwords. On top of this it doesn't have
incorporated salting.

To fix this we introduce an implementation using the Argon2 algorihm,
which is specialized in password-hashing: using proper built-in
salting (which doesn't have to be stored separately), and "Time
Wasting" to make brute forcing harder.