core: [ODIVE-3] better encryption for server messages
This change introduces a better method of encrypting user messages.
By using a random (message-specific) salt we can encrypt the message
then encrypt the salt itself using the API key and then save both
(encrypted message and encrypted salt) in the DB. From there, when
we need to retrieve messages we can decrypt messages one by one.

Note that the launch-scylla-locally.sh changes are already done in
the DB. The changes are necesarry for the migration to the new
encryption method:
- added new column *encrypted* (null on all entries before this
commit)
- added new column *salt*