core: /api/create_server to send SID once created
This change makes this api send a proper payload.
This is part of the ongoing work to make the client
more responsive...

This change also makes some small improvement to the
test environment. One bug that @iv-panzer found
was that the api wasn't getting correctly
recompiled each time the test env is restarted.